Moved verse detail pages to a separate package
